Working only with key sizes 4096
It was going to be a nightmare to deal with discrete versions of the keys, so
I'm only going to work with key sizes of 4096 and have it be able to handle
smaller keys.  Yes, some memory is wasted, but really it's not very much.
The convenience is worth it, and makes the code simpler.