Progress on packet type 2 (Signatures)
It looks like packet type 2 is one of the most complicated (if not *the*
most complicated), and I still don't have subkeys.  It also looks like my
original segment definitions were bad on packet 1, so lets hope I corrected
those.

I moved out a packet 1 function to the utilities because it can be used
commonly.