Simplified packet type 2 data retrieve with unbounded strings
See title.  Utilizing unbounded strings over fixed strings allows the subpacket
data to be delivered immediately, other than embedded signatures.  The latter
still uses the "flag" techique.

Also, the notational data is returned in a fixed array of 10 notes.  If there
are more than this, the remaining are truncated in a similar fashion to how
preferences more than 5 are truncated.  This is considered reasonable by
myself.