Building expat library interface with DOM parsing
I'm working on a simple DOM traversing functionality for XML files.  I really
only need to parse guaranteed good and valid XML files to read the node,
attributes, and values, so this parser is a small subset of a proper library.

Later I'll need to add some tailored functionality to create xml files.