Expat interface already functional
Expat is parsing XML files quite nicely with a DOM interface.  The only major
thing that remains is to convert all the attribute and tag names into
integers.  The plan is to make a class that builds an alphabetical list
appropriate for binary searching.  When a string name is queried, it will
return the index (previous inserting the string to the list if it's the first
instance of it).  This will cut down memory requirements drastically.